requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.10.8
dj-database-url==0.5.0
Django==4.0
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
uuid==1.30
whitenoise==5.3.0
